perbaikan aplikasi mitra, penambahan field foto sisa adonan, part 5

+++ app/Http/Controllers/Api/MitraController.php
@@ -602,42 +602,45 @@ class MitraController extends Controller
         $pengeluaran = MitraOmzetPengeluaranDetail::where('mitra_omzet_pengeluaran_id', $omzet->id)
             ->where('jenis_pengeluaran_mitra_id', $jenis->id)
             ->first();
+        $approved = $pengeluaran ? $pengeluaran->approved : 0;
 
         $deleteName = $pengeluaran->image_nama ? $pengeluaran->image_nama : NULL;
         $deletePath = $pengeluaran->image_lokasi ? $pengeluaran->image_lokasi : NULL;
         $harga = $pengeluaran->harga ? $pengeluaran->harga * $pengeluaran->jumlah : 0;
         $deleteSuccess = false;
 
-        try {
-            $pengeluaran->delete();
-            if ($deleteName && $deletePath) {
-                File::delete(public_path($deletePath) . '/' . $deleteName);
-            }
-            $deleteSuccess = true;
-        } catch (\Illuminate\Database\QueryException $e) {
-            $this->db_switch(1);
+        if ($approved <> 1) {
+            try {
+                $pengeluaran->delete();
+                if ($deleteName && $deletePath) {
+                    File::delete(public_path($deletePath) . '/' . $deleteName);
+                }
+                $deleteSuccess = true;
+            } catch (\Illuminate\Database\QueryException $e) {
+                $this->db_switch(1);
 
-            return response()->json([
-                'status' => 'error',
-                'message' => $e->getMessage(),
-            ]);
-        }
-
-        if ($deleteSuccess) {
-            $date = Carbon::parse($data['tanggal']);
-            $week = $date->isoWeek();
-            $year = $date->isoWeekYear();
-            $yearWeek = $year . str($week)->padLeft(2, '0');
-
-            $kasbon = MitraKasbon::where('isactive', 1)
-                ->where('user_id', $data['id'])
-                ->where('minggu', $yearWeek)
-                ->first();
-
-            if ($kasbon && $jenis->nama == 'Kasbon') {
-                $kasbon->update([
-                    'sisa_plafon' => $kasbon->sisa_plafon + $harga,
+                return response()->json([
+                    'status' => 'error',
+                    'message' => $e->getMessage(),
                 ]);
+            }
+
+            if ($deleteSuccess) {
+                $date = Carbon::parse($data['tanggal']);
+                $week = $date->isoWeek();
+                $year = $date->isoWeekYear();
+                $yearWeek = $year . str($week)->padLeft(2, '0');
+
+                $kasbon = MitraKasbon::where('isactive', 1)
+                    ->where('user_id', $data['id'])
+                    ->where('minggu', $yearWeek)
+                    ->first();
+
+                if ($kasbon && $jenis->nama == 'Kasbon') {
+                    $kasbon->update([
+                        'sisa_plafon' => $kasbon->sisa_plafon + $harga,
+                    ]);
+                }
             }
         }
